refactoring & remove fetch with timeout
